// types/diem.ts
export interface IDiemSo {
  stt: string;
  maHP: string;
  tenHP: string;
  soTC: string;
  danhGia: string;
  chuyenCan: string;
  thi: string;
  tongKet: string;
  diemChu: string;
  [key: string]: number | string;
}

export interface IDiemTongKet {
  namHoc: string;
  hocKy: string;
  TBTL10: string;
  TBTL4: string;
  TC: string;
  TBC10: string;
  TBC4: string;
}